+++ govc/role/remove.go
@@ -1,5 +1,5 @@
 /*
-Copyright (c) 2015 VMware, Inc. All Rights Reserved.
+Copyright (c) 2016 VMware, Inc. All Rights Reserved.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,34 +14,31 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package permissions
+package role
 
 import (
 	"context"
 	"flag"
 
 	"github.com/vmware/govmomi/govc/cli"
-	"github.com/vmware/govmomi/vim25/types"
+	"github.com/vmware/govmomi/govc/permissions"
 )
 
 type remove struct {
-	*PermissionFlag
+	*permissions.PermissionFlag
 
-	types.Permission
-
-	role string
+	force bool
 }
 
 func init() {
-	cli.Register("permissions.remove", &remove{})
+	cli.Register("role.remove", &remove{})
 }
 
 func (cmd *remove) Register(ctx context.Context, f *flag.FlagSet) {
-	cmd.PermissionFlag, ctx = NewPermissionFlag(ctx)
+	cmd.PermissionFlag, ctx = permissions.NewPermissionFlag(ctx)
 	cmd.PermissionFlag.Register(ctx, f)
 
-	f.StringVar(&cmd.Principal, "principal", "", "User or group for which the permission is defined")
-	f.BoolVar(&cmd.Group, "group", false, "True, if principal refers to a group name; false, for a user name")
+	f.BoolVar(&cmd.force, "force", false, "Force removal if role is in use")
 }
 
 func (cmd *remove) Process(ctx context.Context) error {
@@ -52,21 +49,20 @@ func (cmd *remove) Process(ctx context.Context) error {
 }
 
 func (cmd *remove) Usage() string {
-	return "[PATH]..."
+	return "NAME"
 }
 
 func (cmd *remove) Description() string {
-	return `Removes a permission rule from managed entities.
+	return `Remove authorization role.
 
 Examples:
-  govc permissions.remove -principal root
-  govc permissions.remove -principal $USER@vsphere.local -role Admin /dc1/host/cluster1`
+  govc role.remove MyRole
+  govc role.remove MyRole -force`
 }
 
 func (cmd *remove) Run(ctx context.Context, f *flag.FlagSet) error {
-	refs, err := cmd.ManagedObjects(ctx, f.Args())
-	if err != nil {
-		return err
+	if f.NArg() != 1 {
+		return flag.ErrHelp
 	}
 
 	m, err := cmd.Manager(ctx)
@@ -74,12 +70,10 @@ func (cmd *remove) Run(ctx context.Context, f *flag.FlagSet) error {
 		return err
 	}
 
-	for _, ref := range refs {
-		err = m.RemoveEntityPermission(ctx, ref, cmd.Principal, cmd.Group)
-		if err != nil {
-			return err
-		}
+	role, err := cmd.Role(f.Arg(0))
+	if err != nil {
+		return err
 	}
 
-	return nil
+	return m.RemoveRole(ctx, role.RoleId, !cmd.force)
 }