+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # changelog
 
+### unreleased
+
+* Add object.AuthorizationManager methods: RetrieveRolePermissions, RetrieveAllPermissions, AddRole, RemoveRole, UpdateRole
+
 ### 0.11.3 (2016-11-08)
 
 * Allow DatastoreFile.Follow reader to drain current body after stopping